Server/requirements.txt
absl-py==0.9.0
astor==0.8.1
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.24
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
dnspython==2.0.0
Flask==1.0.2
Flask-Mail==0.9.1
Flask-SocketIO==3.1.2
Flask-Uploads==0.2.1
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
msrest==0.6.11
mtcnn==0.1.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
pandas==1.1.3
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.2
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.8
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.11.2
